# fix glaslos/ssdeep#5