I looked at the c implementation and ran both on few files and found the mismatches.
fixes #5