@glaslos 2017-12-06
Nice catch! Can you add the result of `make benchcmp` as a comment here? I don't expect any significant performance changes, but just to be save.