@glaslos 2017-12-06
Actually I just went ahead and made one :)
```
benchmark                  old ns/op     new ns/op     delta
BenchmarkDistance-4        8610          8571          -0.45%
BenchmarkRollingHash-4     2.67          7.04          +163.67%
BenchmarkSumHash-4         29.5          29.5          +0.00%
BenchmarkBlockSize-4       8.97          8.76          -2.34%
```
Do you want to spot your issue yourself or shall I spoil the fun?
Btw, `rollHash` is called very often, so it's performance is critical. Critical enough to write less OO code :)