@davidt99 2017-12-06
It comes down to whether use `rollSum` as a function or not (inline).
I thought about it when I made the change (didn't run the benchmark though) but decided to go with the function because:
1. The function is been used in another place.
2. I wanted to be close to the C code and their `roll_sum` is a separated function from `roll_hash`.
3. The change is not significant (5 nanosecond) and doesn't affect in the overall process performance.
4. In general, when optimizing a code, probably the last resort is to convert a function to an inline function.

Anyway, I move the `rollSum` outside of the `rollHash` so the benchmark stays the same.
In the future, I plan to make more optimizations, focusing on working with buffers instead of reading all the file into memory.